RelationRecognition/requirements.txt
backports.functools-lru-cache==1.6.1
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.1.0
matplotlib==2.2.5
networkx==2.2
numpy==1.16.6
opencv-python==4.1.0.25
Pillow==6.2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.0.3
PyYAML==5.3.1
scikit-image==0.14.5
scipy==1.2.3
six==1.15.0
subprocess32==3.5.4
torch==0.4.0
tqdm==4.46.0
